requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
jinja2==3.1.3
pydantic==2.5.3
gspread==6.0.0
google-auth==2.27.0
